==> Stockton/01_Data Extraction I.ipynb <==
import csv

# Train DataSet

# March
stations = ['1096710','1096310','1007410','1021310','1021410','1021510','1016610','1027410','10115310','1022010','1022210']

with open('03March.csv', 'w') as new_file:
    csv_writer = csv.writer(new_file)
    
    for i in range(1,32):
        with open('DataSet/Traffic/03March/d10_text_station_5min_2019_03_{}.txt'.
                  format(str(i).zfill(2)), 'r') as csv_file:
            csv_reader = csv.reader(csv_file)
    
            for line in csv_reader:
                if line[1] in stations:
                    csv_writer.writerow(line)

# April
stations = ['1096710','1096310','1007410','1021310','1021410','1021510','1016610','1027410','10115310','1022010','1022210']

with open('04April.csv', 'w') as new_file:
    csv_writer = csv.writer(new_file)
    
    for i in range(1,31):
        with open('DataSet/Traffic/04April/d10_text_station_5min_2019_04_{}.txt'.
                  format(str(i).zfill(2)), 'r') as csv_file:
            csv_reader = csv.reader(csv_file)
    
            for line in csv_reader:
                if line[1] in stations:
                    csv_writer.writerow(line)

# May
stations = ['1096710','1096310','1007410','1021310','1021410','1021510','1016610','1027410','10115310','1022010','1022210']

with open('05May.csv', 'w') as new_file:
    csv_writer = csv.writer(new_file)
    
    for i in range(1,32):
        with open('DataSet/Traffic/05May/d10_text_station_5min_2019_05_{}.txt'.
                  format(str(i).zfill(2)), 'r') as csv_file:
            csv_reader = csv.reader(csv_file)
    
            for line in csv_reader:
                if line[1] in stations:
                    csv_writer.writerow(line)

# June
stations = ['1096710','1096310','1007410','1021310','1021410','1021510','1016610','1027410','10115310','1022010','1022210']

with open('06June.csv', 'w') as new_file:
    csv_writer = csv.writer(new_file)
    
    for i in range(1,31):
        with open('DataSet/Traffic/06June/d10_text_station_5min_2019_06_{}.txt'.
                  format(str(i).zfill(2)), 'r') as csv_file:
            csv_reader = csv.reader(csv_file)
    
            for line in csv_reader:
                if line[1] in stations:
                    csv_writer.writerow(line)

# Test DataSet

# Weekday 6th Aug 2019
stations = ['1096710','1096310','1007410','1021310','1021410','1021510','1016610','1027410','10115310','1022010','1022210']

with open('07Weekday.csv', 'w') as new_file:
    csv_writer = csv.writer(new_file)
    
    with open('DataSet/Traffic/Test/d10_text_station_5min_2019_08_06.txt', 'r') as csv_file:
        csv_reader = csv.reader(csv_file)
    
        for line in csv_reader:
            if line[1] in stations:
                csv_writer.writerow(line)

# Weekend 10th Aug 2019
stations = ['1096710','1096310','1007410','1021310','1021410','1021510','1016610','1027410','10115310','1022010','1022210']

with open('08Weekend.csv', 'w') as new_file:
    csv_writer = csv.writer(new_file)
    
    with open('DataSet/Traffic/Test/d10_text_station_5min_2019_08_10.txt', 'r') as csv_file:
        csv_reader = csv.reader(csv_file)
    
        for line in csv_reader:
            if line[1] in stations:
                csv_writer.writerow(line)
