==> Stockton/02_Data Extraction II.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

# Train DataSet

# Read saved csv files
March = pd.read_csv('03March.csv',header=None)
April = pd.read_csv('04April.csv',header=None)
May   = pd.read_csv('05May.csv',header=None)
June  = pd.read_csv('06June.csv',header=None)

# Trim the data (first 17 columns)
March = March.iloc[:,0:17]
April = April.iloc[:,0:17]
May   = May.iloc[:,0:17]
June  = June.iloc[:,0:17]

# Define dataframe header
columns = ['Timestamp','Station','District','Freeway','DoT','Lane Type','Station Length','Samples', '% Obs',
          'Total flow','Avg Occ','Ave Speed','Lane N Samples','Lane N Flow','Lane N Avg Occ','Lane N Avg Speed','Lane N obs']

# Input the header defined above
March.columns = columns
April.columns = columns
May.columns   = columns
June.columns  = columns

May.head()

March.tail()

df = pd.concat([March,April,May,June],ignore_index=True)

len(df)

# Pass a list of column names (the important variables)
df = df[['Timestamp','Station','Total flow']]

df

# Replace Station ID
mymap = {
            1096710  : 'f1',
            1096310  : 'f2',
            1007410  : 'f3',
            1021310  : 'f4',
            1021410  : 'f5',
            1021510  : 'f6',
            1016610  : 'f7',
            1027410  : 'f8',
            10115310 : 'f9',
            1022010  : 'f10',
            1022210  : 'f11'
        }
df['Station'] = df['Station'].map(mymap)

df['Station'].value_counts()

df['Timestamp'].nunique()

# Re-arrange the Dataframe to spatio-temporal format
df = df.pivot(index='Timestamp',columns='Station',values='Total flow')

df.keys()

# f11 is incomplete so drop it
df.drop(columns = 'f11', inplace = True)

df

# Re-arrange spatial columns, so, the data is sequential
df = df[['f1','f2','f3','f4','f5','f6','f7','f8','f9','f10']]
df

df.to_csv('000TrainVal.csv')

# Test DataSet

# Read CSV data
Weekday   = pd.read_csv('07Weekday.csv',header=None)
Weekend   = pd.read_csv('08Weekend.csv',header=None)

# Trim the data (first 17 columns only)
Weekday = Weekday.iloc[:,0:17]
Weekend = Weekend.iloc[:,0:17]

# Rename the columns with the column variable from above
Weekday.columns = columns
Weekend.columns = columns

# Only these three variables are required
Weekday = Weekday[['Timestamp','Station','Total flow']]
Weekend = Weekend[['Timestamp','Station','Total flow']]

# Rename the station IDs
Weekday['Station'] = Weekday['Station'].map(mymap)
Weekend['Station'] = Weekend['Station'].map(mymap)

# Re-arrange the data to time-series format
Weekday = Weekday.pivot(index='Timestamp',columns='Station',values='Total flow')
Weekend = Weekend.pivot(index='Timestamp',columns='Station',values='Total flow')

# Drop the f11 column (unrequired)
Weekday.drop(columns = 'f11', inplace = True)
Weekend.drop(columns = 'f11', inplace = True)

# Re-arrange the station header to be sequential
Weekday = Weekday[['f1','f2','f3','f4','f5','f6','f7','f8','f9','f10']]
Weekend = Weekend[['f1','f2','f3','f4','f5','f6','f7','f8','f9','f10']]

Weekday

Weekend

Weekday.to_csv('00Weekday.csv')
Weekend.to_csv('00Weekend.csv')
