==> Stockton/03d_DPP_Data Scaling.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

# Read saved csv files
TrainVal     = pd.read_csv('01TrainVal.csv')

TrainVal['index'] = pd.to_datetime(TrainVal['index'])
TrainVal.rename(columns = {'index':'Timestamp'}, inplace=True)

TrainVal['Timestamp']

TrainVal

# Split to Training and Test Data

len(TrainVal)

test_percent = 0.11475     # 11.475 % used as test data (final 2 weeks)
len(TrainVal)*test_percent
test_point = np.round(len(TrainVal)*test_percent)
test_ind = int(len(TrainVal) - test_point)

test_ind

train = TrainVal.iloc[:test_ind].drop('Timestamp', axis=1)      # From 1st element to 31575
test = TrainVal.iloc[test_ind:].drop('Timestamp', axis=1)       # From 31575 to end

test.head()

test_percen = 0.12962     # 11.475 % used as test data (final 2 weeks)
len(train)*test_percen
test_p = np.round(len(train)*test_percen)
test_ind = int(len(train) - test_p)

test_ind

t = train.iloc[:test_ind]      # From 1st element to 31575

t

TrainVal.iloc[27071]

# Feature Scaling

# The scaler - for min-max scaling
from sklearn.preprocessing import MinMaxScaler

The MinMaxScaler from scikit-learn subtracts the minimum value and re-scales the the value range, rendering the variable values between 0 and 1. 

# set up the scaler
scaler = MinMaxScaler()

# fit the scaler to the train set, it will learn the parameters
scaler.fit(train)

# transform train and test sets
train_scaled = scaler.transform(train)
test_scaled = scaler.transform(test)

# the scaler stores the maximum values of the features, learned from train set
scaler.data_max_

# the scaler stores the minimum values of the features, learned from train set
scaler.data_min_

l = ['f1','f2','f3','f4','f5','f6','f7']

train_scaled = pd.DataFrame(train_scaled, columns=l)
test_scaled = pd.DataFrame(test_scaled, columns=l)

train_scaled['Timestamp'] = TrainVal['Timestamp'].iloc[:test_ind]
test_scaled['Timestamp'] = TrainVal.iloc[test_ind:].reset_index()['Timestamp']

train_scaled

test_scaled

type(test_scaled.iloc[0]['Timestamp'])

# Save Data and Scaler

import joblib

scaler_filename = "scaler.save"
joblib.dump(scaler, scaler_filename) 

train.to_csv('01train.csv')
test.to_csv('01test.csv')

train_scaled.to_csv('01train_scaled.csv')
test_scaled.to_csv('01test_scaled.csv')
