==> Stockton/03f_DPP_PreviousWeek.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import joblib

Ext             = pd.DataFrame(joblib.load("Extension.save"))
TrainVal_scaled = pd.read_csv('01train_scaled.csv').drop(['Timestamp','Unnamed: 0'], axis=1)
Test_scaled     = pd.read_csv('01test_scaled.csv').drop(['Timestamp','Unnamed: 0'], axis=1)

columns = ['f1', 'f2', 'f3', 'f4', 'f5', 'f6', 'f7']
Ext.columns = columns

Ext

TrainVal_scaled

Test_scaled

### Total Data

df = pd.concat([Ext, TrainVal_scaled, Test_scaled], ignore_index=True)

df

date_range = pd.date_range('2019-02-14 00:00', '2019-06-30 23:59', freq='5Min')
myser = pd.Series(data=date_range, name='Timestamp')

len(myser)

df.drop(columns = ['f1', 'f2', 'f3', 'f4', 'f6', 'f7'], inplace=True);
df['Timestamp'] = myser

# Re-arrange spatial columns, so, the data is sequential
df = df[['Timestamp','f5']]

df

df.iloc[4320]          # First sample

df.iloc[2304]          # -1 week (-2016 samples)

df.iloc[288]          # -2 week (-4032 samples)

# Generate Weekly Samples

a = np.zeros((2, 7))
b = np.zeros((2, 7))
x = 4320                           # Start

for i in range(len(df.iloc[x:])):
    for j in range(2):
        for k in range(7):
            if j == 0:
                a[j][k] = df.iloc[x-2019+i+k].drop(['Timestamp'])
            else:
                a[j][k] = df.iloc[x-4035+i+k].drop(['Timestamp'])
    b = np.append(b, a, axis=0)

b = np.delete(b, np.s_[0:2], 0).reshape(-1, 2, 7)

b[0]

b[1]

b[0].flatten()

a = b.flatten().reshape(-1,14)

a[0]

len(a) == len(TrainVal_scaled) + len(Test_scaled)

### Split into corresponding test and train

train = a[:TrainVal_scaled.shape[0]]

train.shape

TrainVal_scaled.shape

test = a[TrainVal_scaled.shape[0]:]

test.shape

Test_scaled.shape

#### Confirmation

test[0]

TrainVal_scaled.iloc[-2016]['f5']  

# Save Data

joblib.dump(train, "002weeks_train.save");
joblib.dump(test, "002weeks_test.save");
