==> Stockton/04_Sample Generators_I_Deep-Multi_Output.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

#### Read training data from csv files

Train     = pd.read_csv('01train_scaled.csv', index_col=0, parse_dates=[8])
Test      = pd.read_csv('01test_scaled.csv', index_col=0, parse_dates=[8])

Test

type(Test.iloc[0]['Timestamp'])

train = Train.drop('Timestamp', axis=1)
test = Test.drop('Timestamp', axis=1)

train

# 5-minute (1 step ahead) foreast

## Train

length = 15
step = 1
k = -1

a = np.zeros((15, 7))
b = np.zeros((7,))
x = np.zeros((15, 7))
y = np.zeros((7,))


for i in range(length, (len(train)-step+1)):
    b = np.asarray(train.iloc[i+step-1])
    k = k+1
    for j in range(length):
        if j+k <= (len(train) + length - step):
            a[j] = train.iloc[j+k]
            
    x = np.append(x, a, axis=0)
    y = np.append(y, b, axis=0)
    
x = np.delete(x, np.s_[0:15], 0).reshape(-1, 15, 7)
y = np.delete(y, np.s_[0:7], 0).reshape(-1, 7)

x.shape

y.shape

train.tail(16)

y[-1]

# Checked as correct

# Save 

np.savez_compressed("Deep_train_5_multi", x=x , y=y)

## Test

length = 15
step = 1
k = -1

a = np.zeros((15, 7))
b = np.zeros((7,))
x = np.zeros((15, 7))
y = np.zeros((7,))


for i in range(length, (len(test)-step+1)):
    b = np.asarray(test.iloc[i+step-1])
    k = k+1
    for j in range(length):
        if j+k <= (len(test) + length - step):
            a[j] = test.iloc[j+k]
            
    x = np.append(x, a, axis=0)
    y = np.append(y, b, axis=0)
    
x = np.delete(x, np.s_[0:15], 0).reshape(-1, 15, 7)
y = np.delete(y, np.s_[0:7], 0).reshape(-1, 7)

x.shape

y.shape

test.tail()

y[-1]

# Checked as correct

# Save 

np.savez_compressed("Deep_test_5_multi", x=x , y=y)

# 15-minute (3 step ahead) forecast

## Train

length = 15
step = 3
k = -1

a = np.zeros((15, 7))
b = np.zeros((7,))
x = np.zeros((15, 7))
y = np.zeros((7,))


for i in range(length, (len(train)-step+1)):
    b = np.asarray(train.iloc[i+step-1])
    k = k+1
    for j in range(length):
        if j+k <= (len(train) + length - step):
            a[j] = train.iloc[j+k]
            
    x = np.append(x, a, axis=0)
    y = np.append(y, b, axis=0)
    
x = np.delete(x, np.s_[0:15], 0).reshape(-1, 15, 7)
y = np.delete(y, np.s_[0:7], 0).reshape(-1, 7)

x.shape

y.shape

train.tail(16)

y[-1]

# Checked as correct

# Save 

np.savez_compressed("Deep_train_15_multi", x=x , y=y)

## Test

length = 15
step = 3
k = -1

a = np.zeros((15, 7))
b = np.zeros((7,))
x = np.zeros((15, 7))
y = np.zeros((7,))


for i in range(length, (len(test)-step+1)):
    b = np.asarray(test.iloc[i+step-1])
    k = k+1
    for j in range(length):
        if j+k <= (len(test) + length - step):
            a[j] = test.iloc[j+k]
            
    x = np.append(x, a, axis=0)
    y = np.append(y, b, axis=0)
    
x = np.delete(x, np.s_[0:15], 0).reshape(-1, 15, 7)
y = np.delete(y, np.s_[0:7], 0).reshape(-1, 7)

x.shape

y.shape

test.tail()

y[-1]

# Checked as correct

# Save 

np.savez_compressed("Deep_test_15_multi", x=x , y=y)

# 30-minute (6 step ahead) forecast

## Train

length = 15
step = 6
k = -1

a = np.zeros((15, 7))
b = np.zeros((7,))
x = np.zeros((15, 7))
y = np.zeros((7,))


for i in range(length, (len(train)-step+1)):
    b = np.asarray(train.iloc[i+step-1])
    k = k+1
    for j in range(length):
        if j+k <= (len(train) + length - step):
            a[j] = train.iloc[j+k]
            
    x = np.append(x, a, axis=0)
    y = np.append(y, b, axis=0)
    
x = np.delete(x, np.s_[0:15], 0).reshape(-1, 15, 7)
y = np.delete(y, np.s_[0:7], 0).reshape(-1, 7)

x.shape

y.shape

train.tail()

y[-1]

# Checked as correct

# Save 

np.savez_compressed("Deep_train_30_multi", x=x , y=y)

## Test

length = 15
step = 6
k = -1

a = np.zeros((15, 7))
b = np.zeros((7,))
x = np.zeros((15, 7))
y = np.zeros((7,))


for i in range(length, (len(test)-step+1)):
    b = np.asarray(test.iloc[i+step-1])
    k = k+1
    for j in range(length):
        if j+k <= (len(test) + length - step):
            a[j] = test.iloc[j+k]
            
    x = np.append(x, a, axis=0)
    y = np.append(y, b, axis=0)
    
x = np.delete(x, np.s_[0:15], 0).reshape(-1, 15, 7)
y = np.delete(y, np.s_[0:7], 0).reshape(-1, 7)

x.shape

y.shape

test.tail()

y[-1]

# Checked as correct

# Save 

np.savez_compressed("Deep_test_30_multi", x=x , y=y)

# 60-minute (12 step ahead) forecast

## Train

length = 15
step = 12
k = -1

a = np.zeros((15, 7))
b = np.zeros((7,))
x = np.zeros((15, 7))
y = np.zeros((7,))


for i in range(length, (len(train)-step+1)):
    b = np.asarray(train.iloc[i+step-1])
    k = k+1
    for j in range(length):
        if j+k <= (len(train) + length - step):
            a[j] = train.iloc[j+k]
            
    x = np.append(x, a, axis=0)
    y = np.append(y, b, axis=0)
    
x = np.delete(x, np.s_[0:15], 0).reshape(-1, 15, 7)
y = np.delete(y, np.s_[0:7], 0).reshape(-1, 7)

x.shape

y.shape

train.tail()

y[-1]

# Checked as correct

# Save 

np.savez_compressed("Deep_train_60_multi", x=x , y=y)

## Test

length = 15
step = 12
k = -1

a = np.zeros((15, 7))
b = np.zeros((7,))
x = np.zeros((15, 7))
y = np.zeros((7,))


for i in range(length, (len(test)-step+1)):
    b = np.asarray(test.iloc[i+step-1])
    k = k+1
    for j in range(length):
        if j+k <= (len(test) + length - step):
            a[j] = test.iloc[j+k]
            
    x = np.append(x, a, axis=0)
    y = np.append(y, b, axis=0)
    
x = np.delete(x, np.s_[0:15], 0).reshape(-1, 15, 7)
y = np.delete(y, np.s_[0:7], 0).reshape(-1, 7)

x.shape

y.shape

test.tail()

y[-1]

# Checked as correct

# Save 

np.savez_compressed("Deep_test_60_multi", x=x , y=y)
